apex4: add GIF upload capture
For some reason Flydigi Space 3.0 doesn't allow changing the GIF
unless you switch to xinput mode, so this capture is the apex 4
communication when in xinput and with the upload of a new GIF to
the screen.

The upload of a 1.927 KB GIF takes some time (~ around 2 minutes)
so it explains the size of the capture.